Given I've renamed the test name to be safe the filename and switch name also needs to be changed.
